CHS version of the KAOS Architecture
This KAOS Baseline is useful for building a Kernel on computers with older Hard Drives.
Since the basic KAOS Horizons architecture is intended to run on a USB Flash Drive or modern Solid State drive, this is merely a side hustle to leave no stone unturned. Gotsta be thorough. Gotta learn all this stuff to get this far. Therefore we are compelled to share our data on the near impossible journey that got us thus far.